Refactor PriceCalculator -> PurchaseForm
