+++ apps/store/src/blocks/PriceCalculatorBlock.tsx
@@ -1,11 +1,11 @@
 import { storyblokEditable } from '@storyblok/react'
-import { PriceCalculator } from '@/components/ProductPage/PriceCalculator/PriceCalculator'
+import { PurchaseForm } from '@/components/ProductPage/PurchaseForm/PurchaseForm'
 import { SbBaseBlockProps } from '@/services/storyblok/storyblok'
 
 type Props = SbBaseBlockProps<unknown>
 
 export const PriceCalculatorBlock = ({ blok }: Props) => {
-  return <PriceCalculator {...storyblokEditable(blok)} />
+  return <PurchaseForm {...storyblokEditable(blok)} />
 }
 
 PriceCalculatorBlock.blockName = 'priceCalculator'
